Modify each section of content to have a consistent max width
Added containers to the wrap the content of each section and applied a
max width to each of them so that the webpage will have a consistent
width regardless of how much a user may zoom out of it. This will
prevent each section of content from expanding to the full width of
their parent when zooming out, as this was making it look less
professional and was inconsistent from one section to another.