Adjust look of website for smaller screens
Added side margins to each section of the website so that there will
be space between the content and the edge of the screen, resulting
in a better view of the website on smaller screen widths.

Also adjusted the distribution of content on the information section
so that the information cards will line up at the center instead of
at the flex start whenever they wrap to a new line.